Removed println and added default and added install.kt
