+++ plugins/server/org.mongodb/mongodb/2.2/Databases.kt
@@ -6,42 +6,8 @@ import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
-fun Application.configureDatabases() {
-    val mongoDatabase = connectToMongoDB()
-    val carService = CarService(mongoDatabase)
-    routing {
-        // Create car
-        post("/cars") {
-            val car = call.receive<Car>()
-            val id = carService.create(car)
-            println("id : $id")
-            call.respond(HttpStatusCode.Created, id)
-        }
-        // Read car
-        get("/cars/{id}") {
-            val id = call.parameters["id"] ?: throw IllegalArgumentException("No ID found")
-            carService.read(id)?.let { car ->
-                call.respond(car)
-            } ?: call.respond(HttpStatusCode.NotFound)
-        }
-        // Update car
-        put("/cars/{id}") {
-            val id = call.parameters["id"] ?: throw IllegalArgumentException("No ID found")
-            val car = call.receive<Car>()
-            carService.update(id, car)?.let {
-                call.respond(HttpStatusCode.OK)
-            } ?: call.respond(HttpStatusCode.NotFound)
-        }
-        // Delete car
-        delete("/cars/{id}") {
-            val id = call.parameters["id"] ?: throw IllegalArgumentException("No ID found")
-            carService.delete(id)?.let {
-                call.respond(HttpStatusCode.OK)
-            } ?: call.respond(HttpStatusCode.NotFound)
-        }
-    }
-}
-
+val mongoDatabase = connectToMongoDB()
+val carService = CarService(mongoDatabase)
 /**
  * Establishes connection with a MongoDB database.
  *
