+++ plugins/server/org.mongodb/mongodb/2.2/install.kt
@@ -1,0 +1,39 @@
+import io.ktor.http.*
+import io.ktor.server.application.*
+import io.ktor.server.request.*
+import io.ktor.server.response.*
+import java.sql.*
+import kotlinx.coroutines.*
+
+public fun Application.configureDatabases() {
+    routing {
+        // Create car
+        post("/cars") {
+            val car = call.receive<Car>()
+            val id = carService.create(car)
+            call.respond(HttpStatusCode.Created, id)
+        }
+        // Read car
+        get("/cars/{id}") {
+            val id = call.parameters["id"] ?: throw IllegalArgumentException("No ID found")
+            carService.read(id)?.let { car ->
+                call.respond(car)
+            } ?: call.respond(HttpStatusCode.NotFound)
+        }
+        // Update car
+        put("/cars/{id}") {
+            val id = call.parameters["id"] ?: throw IllegalArgumentException("No ID found")
+            val car = call.receive<Car>()
+            carService.update(id, car)?.let {
+                call.respond(HttpStatusCode.OK)
+            } ?: call.respond(HttpStatusCode.NotFound)
+        }
+        // Delete car
+        delete("/cars/{id}") {
+            val id = call.parameters["id"] ?: throw IllegalArgumentException("No ID found")
+            carService.delete(id)?.let {
+                call.respond(HttpStatusCode.OK)
+            } ?: call.respond(HttpStatusCode.NotFound)
+        }
+    }
+}
