+++ plugins/server/org.mongodb/mongodb/2.2/manifest.ktor.yaml
@@ -8,6 +8,6 @@ prerequisites:
   - kotlinx-serialization
 documentation: documentation.md
 installation:
+  default: install.kt
   outside_app: Databases.kt
   source_file_kt: CarsSchema.kt
-
